moved api tests to test_api
