+++ eltelive-new/package.json
@@ -8,7 +8,7 @@
     "lint": "vue-cli-service lint",
     "server": "node server/server.js",
     "watch": "nodemon server/server.js",
-    "test": "mocha server/test/test.js --exit"
+    "test:api": "mocha server/test/test_api/test.js --exit"
   },
   "dependencies": {
     "@vue/eslint-plugin": "^4.2.0",
