+++ eltelive-new/server/test/test_api/delete_key.js
@@ -1,17 +1,17 @@
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 const jwt = require('jsonwebtoken');
-const mongoose = require('../db_connections/test');
-const User = require('../model/user');
-const server = require('../server');
+const mongoose = require('../../db_connections/test');
+const User = require('../../model/user');
+const server = require('../../server');
 const temp_data = require('./temp_data')
 
 const should = chai.should();
 chai.use(chaiHttp);
 
-describe('/GET get_users', async () => {
-    it('should return the details of all of the users, in case of the admin being signed in', async () => {
-        const user = await User.findOne({ email: temp_data.ADMIN_USER.email }).lean()
+describe('/DELETE delete_key', async () => {
+    it('should delete the stream key of the logged-in user', async () => {
+        const user = await User.findOne({ email: temp_data.TEST1_USER.email }).lean()
         const token = jwt.sign(
             {
                 id: user._id,
@@ -20,20 +20,19 @@ describe('/GET get_users', async () => {
             process.env.JWT_SECRET
         )
         chai.request(server)
-            .get('/api/get_users')
+            .delete('/api/delete_key')
             .set({ "Authorization": `Bearer ${token}` })
             .end((err, res) => {
                 res.body.should.be.a('object');
                 res.should.have.status(200);
-                res.body.title.should.be.eql('Users details are retrieved successfully');
-                res.body.should.have.property('users');
+                res.body.title.should.be.eql('Stream key deleted successfully');
             });
     })
 
-    it('should return "JWT Token not provided" error, if the token is missing', async () => {
+    it('should return "JWT Token not provided" error', async () => {
         const token = temp_data.EMPTY_STRING
         chai.request(server)
-            .get('/api/get_users')
+            .delete('/api/delete_key')
             .set({ "Authorization": `Bearer ${token}` })
             .end((err, res) => {
                 res.body.should.be.a('object');
@@ -42,10 +41,10 @@ describe('/GET get_users', async () => {
             });
     })
 
-    it('should return "Unexpected error" error, if the token is not in the correct format', async () => {
+    it('should return "Unexpected error" error', async () => {
         const token = temp_data.DUMMY_STRING
         chai.request(server)
-            .get('/api/get_users')
+            .delete('/api/delete_key')
             .set({ "Authorization": `Bearer ${token}` })
             .end((err, res) => {
                 res.body.should.be.a('object');
@@ -63,31 +62,12 @@ describe('/GET get_users', async () => {
             process.env.JWT_SECRET
         )
         chai.request(server)
-            .get('/api/get_users')
+            .delete('/api/delete_key')
             .set({ "Authorization": `Bearer ${token}` })
             .end((err, res) => {
                 res.body.should.be.a('object');
                 res.should.have.status(404);
                 res.body.title.should.be.eql('User with this token does not exist');
-            });
-    })
-
-    it('should return "Only the admin can get the list of users" error, in case of someone different than the admin being signed in', async () => {
-        const user = await User.findOne({ email: temp_data.TEST1_USER.email }).lean()
-        const token = jwt.sign(
-            {
-                id: user._id,
-                email: user.email
-            },
-            process.env.JWT_SECRET
-        )
-        chai.request(server)
-            .get('/api/get_users')
-            .set({ "Authorization": `Bearer ${token}` })
-            .end((err, res) => {
-                res.body.should.be.a('object');
-                res.should.have.status(403);
-                res.body.title.should.be.eql('Only the admin can get the list of users');
             });
     })
 });
