+++ eltelive-new/server/test/test_api/register.js
@@ -1,9 +1,9 @@
 const chai = require('chai');
 const chaiHttp = require('chai-http');
 const jwt = require('jsonwebtoken');
-const mongoose = require('../db_connections/test');
-const User = require('../model/user');
-const server = require('../server');
+const mongoose = require('../../db_connections/test');
+const User = require('../../model/user');
+const server = require('../../server');
 const temp_data = require('./temp_data')
 
 const should = chai.should();
